Bug #1321803 soc: sprd: iommu: fix incorrect variable type for map_cnt
[root cause  ] iommu map counter overflow
[changes     ] use u32 for map_cnt
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:sharkl5pro
[test case]:vdsp stress
[reviewers   ] Wenan Hu
[change_type ] feature_bugfix
[tag_product ] specific

Fix incorrect variable type for map_cnt.

+++ drivers/soc/sprd/iommu/drv/cll/sprd_iommuvau_cll.h
@@ -38,7 +38,7 @@ struct sprd_iommuvau_priv {
 
 	u8 ram_clk_div;/*Clock divisor*/
 
-	u8 map_cnt;
+	u32 map_cnt;
 	enum sprd_iommu_type iommu_type;
 	enum IOMMU_ID iommu_id;
 	int chip;
